chore(lists): rework list updating logic
