+++ src/pages/ProjectListsPage/hooks/useUpdateList.ts
@@ -1,13 +1,19 @@
 import { useState } from 'react'
 import { LIST_CATEGORY_ATTRIBUTE, useListsDataContext } from '../context/ListsDataContext'
 import { ListsContextType } from '../context'
-import { usePatchAttributeConfigMutation, type EntityListPatchModel } from '@shared/api'
+import {
+  usePatchAttributeConfigMutation,
+  useUpdateEntityListMutation,
+  type EntityListPatchModel,
+} from '@shared/api'
 import { toast } from 'react-toastify'
 import { useCallback } from 'react'
+import { kebabCase } from 'lodash'
 
 export interface UseUpdateListProps {
   setRowSelection: ListsContextType['setRowSelection']
   onUpdateList: (listId: string, list: EntityListPatchModel) => Promise<void>
+  projectName: string
 }
 
 export interface UseUpdateListReturn {
@@ -15,10 +21,12 @@ export interface UseUpdateListReturn {
   openRenameList: (listId: string) => void
   closeRenameList: () => void
   submitRenameList: (value: string) => Promise<void>
+  setListsCategory: (listIds: string[], category: string | null) => Promise<void>
+  createAndAssignCategory: (listIds: string[], categoryLabel: string) => Promise<void>
 }
 
-const useUpdateList = ({ setRowSelection, onUpdateList }: UseUpdateListProps) => {
-  const { listsData, categoryAttribute } = useListsDataContext()
+const useUpdateList = ({ setRowSelection, onUpdateList, projectName }: UseUpdateListProps) => {
+  const { listsData, categoryAttribute, categories } = useListsDataContext()
   const [renamingList, setRenamingList] = useState<UseUpdateListReturn['renamingList']>(null)
 
   const openRenameList: UseUpdateListReturn['openRenameList'] = useCallback(
@@ -47,6 +55,8 @@ const useUpdateList = ({ setRowSelection, onUpdateList }: UseUpdateListProps) =>
   }, [])
 
   const [updateAttribute] = usePatchAttributeConfigMutation()
+  const [updateList] = useUpdateEntityListMutation()
+
   const submitRenameList: UseUpdateListReturn['submitRenameList'] = useCallback(
     async (label) => {
       if (!renamingList) return Promise.reject()
@@ -96,7 +106,82 @@ const useUpdateList = ({ setRowSelection, onUpdateList }: UseUpdateListProps) =>
         throw error
       }
     },
-    [renamingList, onUpdateList, closeRenameList],
+    [renamingList, onUpdateList, closeRenameList, categoryAttribute, updateAttribute],
+  )
+
+  const setListsCategory = useCallback(
+    async (listIds: string[], category: string | null) => {
+      try {
+        const updatePromises = listIds.map((listId) =>
+          updateList({
+            listId,
+            projectName,
+            entityListPatchModel: {
+              attrib: {
+                entityListCategory: category,
+              },
+            },
+          }).unwrap(),
+        )
+
+        await Promise.all(updatePromises)
+
+        const message = category
+          ? `Category set to "${category}" for ${listIds.length} list${
+              listIds.length === 1 ? '' : 's'
+            }`
+          : `Category removed from ${listIds.length} list${listIds.length === 1 ? '' : 's'}`
+        toast.success(message)
+      } catch (error: any) {
+        console.error('Failed to update categories:', error)
+        toast.error(`Failed to update categories: ${error.data?.detail || error.message}`)
+        throw error
+      }
+    },
+    [updateList, projectName],
+  )
+
+  const createAndAssignCategory = useCallback(
+    async (listIds: string[], categoryLabel: string) => {
+      // update the attribute config to add the new category
+
+      // generate category name from label
+      const categoryValue = kebabCase(categoryLabel)
+
+      // Check if category already exists
+      const categoryExists = categories?.some((item) => item.value == categoryValue)
+      if (categoryExists) {
+        const errorMsg = `Category "${categoryLabel}" already exists`
+        console.error(errorMsg)
+        toast.error(errorMsg)
+        return
+      }
+
+      const newEnum = [...categories, { label: categoryLabel, value: categoryValue }]
+
+      try {
+        await updateAttribute({
+          attributeName: LIST_CATEGORY_ATTRIBUTE,
+          attributePatchModel: {
+            data: {
+              enum: newEnum,
+            },
+          },
+        }).unwrap()
+
+        // now assign the new category to the lists
+        await setListsCategory(listIds, categoryValue)
+
+        toast.success(
+          `Category "${categoryLabel}" created and assigned to ${listIds.length} list(s)`,
+        )
+      } catch (error: any) {
+        console.error('Failed to create and assign category:', error)
+        toast.error(`Failed to create and assign category: ${error.data?.detail || error.message}`)
+        throw error
+      }
+    },
+    [categories, setListsCategory, updateAttribute],
   )
 
   return {
@@ -104,6 +189,8 @@ const useUpdateList = ({ setRowSelection, onUpdateList }: UseUpdateListProps) =>
     openRenameList,
     closeRenameList,
     submitRenameList,
+    setListsCategory,
+    createAndAssignCategory,
   }
 }
